requirements.txt
Flask==2.1.1
pymongo==3.12.1
requests==2.26.0
geopy==2.2.0
faker==13.3.0
pyyaml==6.0
numpy==1.21.4
python-dotenv==0.19.2
tensorflow==2.10.0
scikit-learn==1.0.2
opencv-python==4.5.5.64
pillow==8.4.0
aiohttp==3.8.1
tweepy==4.4.0
cryptography==3.4.8
